feat: Add interest rate and loan term to loan tracking
NEW FIELDS: Interest Rate & Loan Term! 📊

Database:
- loan_interest_rate: decimal(5,2) - Annual interest rate percentage
- loan_term_months: integer - Loan term in months
- Both nullable, stored after initial_loan_amount

Component Properties:
- loanInterestRate: string (validated as numeric 0-100)
- loanTermMonths: string (validated as integer 1-600)
- Pre-filled when editing existing loans
- Reset on modal close

Validation:
- Interest rate: 0-100% (optional)
- Loan term: 1-600 months / 1-50 years (optional)
- Initial loan amount: Required

UI Layout:
┌─────────────────────────────────┐
│ Initial Loan Amount *           │
│ [25000.00]                      │
├─────────────────┬───────────────┤
│ Interest Rate % │ Loan Term mo  │
│ [5.99] %        │ [60] months   │
└─────────────────┴───────────────┘

Perfect For Calculating:
- Monthly payment amounts
- Total interest paid over life of loan
- Payoff schedules
- Early payoff savings
- Loan amortization charts

Example Values:
- Auto Loan: $25,000 @ 5.99% for 60 months
- Mortgage: $300,000 @ 3.5% for 360 months
- Personal Loan: $10,000 @ 12.99% for 36 months
- Student Loan: $50,000 @ 4.5% for 120 months

+++ app/Models/Bill.php
@@ -175,7 +175,7 @@ class Bill extends Model
      */
     public function getDaysUntilDueAttribute(): int
     {
-        return now()->diffInDays($this->next_due_date, false);
+        return (int) now()->diffInDays($this->next_due_date, false);
     }
 
     /**
